Move fetching identities to store
