+++ src/frontend/src/lib/flows/authLastUsedFlow.svelte.ts
@@ -2,7 +2,7 @@ import {
   authenticateWithJWT,
   authenticateWithPasskey,
 } from "$lib/utils/authentication";
-import { anonymousActor, canisterConfig, canisterId } from "$lib/globals";
+import { canisterConfig, canisterId } from "$lib/globals";
 import { authenticationStore } from "$lib/stores/authentication.store";
 import {
   lastUsedIdentitiesStore,
@@ -11,35 +11,7 @@ import {
 import { createGoogleRequestConfig, requestJWT } from "$lib/utils/openID";
 import { get } from "svelte/store";
 import { sessionStore } from "$lib/stores/session.store";
-import { isNullish, nonNullish } from "@dfinity/utils";
-import { convertToValidCredentialData } from "$lib/utils/credential-devices";
-
-const fetchIdentityCredentials = async (
-  identityNumber: bigint,
-): Promise<Uint8Array[] | undefined> => {
-  try {
-    const identityCredentials = await anonymousActor.lookup(identityNumber);
-    const validCredentials = identityCredentials
-      .filter((device) => "authentication" in device.purpose)
-      .filter(({ key_type }) => !("browser_storage_key" in key_type))
-      .map(convertToValidCredentialData)
-      .filter(nonNullish);
-
-    if (validCredentials.length > 0) {
-      return validCredentials.map(
-        (credential) => new Uint8Array(credential.credentialId),
-      );
-    }
-
-    return undefined;
-  } catch (error) {
-    console.warn(
-      `Error looking up identity ${identityNumber} credentials:`,
-      error,
-    );
-    return undefined;
-  }
-};
+import { isNullish } from "@dfinity/utils";
 
 export class AuthLastUsedFlow {
   systemOverlay = $state(false);
@@ -49,9 +21,9 @@ export class AuthLastUsedFlow {
     try {
       if ("passkey" in lastUsedIdentity.authMethod) {
         // If there is a problem looking up the credentials, we fallback to the credentialId provided by the lastUsedIdentity
-        const credentialIds = (await fetchIdentityCredentials(
-          lastUsedIdentity.identityNumber,
-        )) ?? [lastUsedIdentity.authMethod.passkey.credentialId];
+        const credentialIds = lastUsedIdentity.credentialIds ?? [
+          lastUsedIdentity.authMethod.passkey.credentialId,
+        ];
         const { identity, identityNumber } = await authenticateWithPasskey({
           canisterId,
           session: get(sessionStore),
